+++ lib/widgets/nutrition/forms.dart
@@ -27,12 +27,10 @@ import 'package:wger/models/nutrition/ingredient.dart';
 import 'package:wger/models/nutrition/log.dart';
 import 'package:wger/models/nutrition/meal.dart';
 import 'package:wger/models/nutrition/meal_item.dart';
-import 'package:wger/models/nutrition/nutritional_goals.dart';
 import 'package:wger/models/nutrition/nutritional_plan.dart';
 import 'package:wger/providers/nutrition.dart';
 import 'package:wger/screens/nutritional_plan_screen.dart';
 import 'package:wger/widgets/nutrition/helpers.dart';
-import 'package:wger/widgets/nutrition/macro_nutrients_table.dart';
 import 'package:wger/widgets/nutrition/nutrition_tiles.dart';
 import 'package:wger/widgets/nutrition/widgets.dart';
 
@@ -472,68 +470,6 @@ class IngredientFormState extends State<IngredientForm> {
       ),
     );
   }
-}
-
-void showIngredientDetails(BuildContext context, int id, {String? image}) {
-  // when loading recently used ingredients, we never get an image :'(
-  // we also don't get an image when querying the API
-  // however, the typeahead suggestion does get an image, so we allow passing it...
-
-  final url = context.read<NutritionPlansProvider>().baseProvider.auth.serverUrl;
-
-  showDialog(
-    context: context,
-    builder: (context) => FutureBuilder<Ingredient>(
-      future: Provider.of<NutritionPlansProvider>(context, listen: false).fetchIngredient(id),
-      builder: (BuildContext context, AsyncSnapshot<Ingredient> snapshot) {
-        Ingredient? ingredient;
-        NutritionalGoals? goals;
-
-        if (snapshot.hasData) {
-          ingredient = snapshot.data!;
-          goals = NutritionalGoals(
-            energy: ingredient.nutritionalValues.energy,
-            protein: ingredient.nutritionalValues.protein,
-            carbohydrates: ingredient.nutritionalValues.carbohydrates,
-            carbohydratesSugar: ingredient.nutritionalValues.carbohydratesSugar,
-            fat: ingredient.nutritionalValues.fat,
-            fatSaturated: ingredient.nutritionalValues.fatSaturated,
-            fiber: ingredient.nutritionalValues.fiber,
-            sodium: ingredient.nutritionalValues.sodium,
-          );
-        }
-        return AlertDialog(
-          title: (snapshot.hasData) ? Text(ingredient!.name) : null,
-          content: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                if (image != null)
-                  CircleAvatar(backgroundImage: NetworkImage(url! + image), radius: 128),
-                if (image != null) const SizedBox(height: 12),
-                if (snapshot.hasError)
-                  Text(
-                    'Ingredient lookup error: ${snapshot.error}',
-                    style: const TextStyle(color: Colors.red),
-                  ),
-                if (!snapshot.hasData && !snapshot.hasError) const CircularProgressIndicator(),
-                if (snapshot.hasData)
-                  ConstrainedBox(
-                    constraints: const BoxConstraints(minWidth: 400),
-                    child: MacronutrientsTable(
-                      nutritionalGoals: goals!,
-                      plannedValuesPercentage: goals.energyPercentage(),
-                      nutritionalGoalsGperKg: null,
-                    ),
-                  ),
-              ],
-            ),
-          ),
-        );
-      },
-    ),
-  );
 }
 
 enum GoalType {
